feat: pricing page to rsc
