+++ www/app/layout.tsx
@@ -1,25 +1,16 @@
 import { Analytics } from '@vercel/analytics/react';
 import { ReactNode } from 'react';
-import { Layout } from '../lib/layouts/Layout';
 import { DESCRIPTION, SHORT_DESCRIPTION } from '../lib/constants';
 import { theme } from '../tailwind.config';
+import { Inter } from 'next/font/google';
+import { Header } from '../lib/components/Header';
+import { Footer } from '../lib/components/Footer';
 import '../styles/globals.css';
 import '../lib/posthog';
 
-type RootLayoutProps = {
-  children: ReactNode;
-};
-
-export default function RootLayout({ children }: RootLayoutProps) {
-  return (
-    <html lang="en">
-      <body>
-        <Layout>{children}</Layout>
-        <Analytics />
-      </body>
-    </html>
-  );
-}
+const inter = Inter({
+  subsets: ['latin'],
+});
 
 export const metadata = {
   description: `${SHORT_DESCRIPTION}. ${DESCRIPTION}`,
@@ -43,3 +34,29 @@ export const metadata = {
     },
   },
 };
+
+type LayoutProps = {
+  children: ReactNode;
+};
+
+export default function RootLayout({ children }: LayoutProps) {
+  return (
+    <html lang="en">
+      <body>
+        <div className={`bg-dark pt-8 ${inter.className}`}>
+          <div
+            className="pointer-events-none absolute top-0 left-0 h-64 w-full"
+            style={{
+              background:
+                'linear-gradient(to bottom, transparent, #050211), repeating-linear-gradient(-45deg, #041F47, #041F47 1px, transparent 1px, transparent 20px)',
+            }}
+          />
+          <Header />
+          <main className="container mx-auto flex min-h-screen flex-col gap-32 px-4 pt-24 md:gap-64">{children}</main>
+          <Footer />
+        </div>
+        <Analytics />
+      </body>
+    </html>
+  );
+}
